requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
curl_cffi==0.13.0
dateparser==1.2.2
distro==1.9.0
feedparser==6.0.12
finnhub-python==2.4.26
frozendict==2.4.7
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.29.0
google-auth==2.48.0
google-genai==1.61.0
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jpype1==1.5.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
multidict==6.7.0
multitasking==0.0.12
narwhals==2.15.0
nba_api==1.11.3
nbainjuries==1.0.0
nltk==3.9.2
numpy==2.4.1
packaging==26.0
pandas==2.3.3
peewee==3.19.0
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.2
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.4
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pygooglenews==0.1.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
streamlit==1.53.1
tabula-py==2.10.0
tenacity==9.1.2
textblob==0.19.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
watchdog==6.0.0
websockets==15.0.1
yarl==1.22.0
yfinance==1.1.0
SQLAlchemy==1.4.0
